+++ apps/html/stories/abstract/abstract.stories.ts
@@ -12,6 +12,9 @@ export default {
   title: 'HTML/Typography/Abstract',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: