+++ apps/html/stories/form-error/form-error.stories.ts
@@ -11,6 +11,9 @@ export default {
   title: 'HTML/Forms/Form Error',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: