+++ apps/html/stories/icon/icon.stories.ts
@@ -8,6 +8,9 @@ export default {
   title: 'HTML/Utilities/Icon',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: