+++ apps/html/stories/label-wrapper/label-wrapper.stories.ts
@@ -15,6 +15,9 @@ export default {
   title: 'HTML/Forms/Label Wrapper',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: