+++ apps/html/stories/landing-lead/landing-lead.stories.ts
@@ -5,6 +5,9 @@ export default {
   title: 'HTML/Layout/Landing Lead',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: