+++ apps/html/stories/link/link.stories.ts
@@ -6,6 +6,9 @@ export default {
   title: 'HTML/Controls/Link',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: