+++ apps/html/stories/split-surface/split-surface.stories.ts
@@ -5,6 +5,9 @@ export default {
   title: 'HTML/Layout/Split Surface',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: