+++ apps/html/stories/title/title.stories.ts
@@ -13,6 +13,9 @@ export default {
   title: 'HTML/Typography/Title',
   tags: ['autodocs'],
   parameters: {
+    c1Meta: {
+      classification: ['css-only'],
+    },
     docs: {
       description: {
         component: