+++ apps/react/.storybook/preview.ts
@@ -37,8 +37,12 @@ import '../../../packages/ui/dist/css/components/image.css';
 import '../../../packages/ui/dist/css/components/prose.css';
 import '../../../packages/ui/dist/css/components/list.css';
 import '../../../packages/ui/dist/css/components/icon.css';
+import { c1DocsPage } from '../../storybook/docs/c1-docs-page';
 
 export const parameters = {
+  docs: {
+    page: c1DocsPage,
+  },
   controls: {
     expanded: true,
     matchers: {