+++ apps/react/src/stories/accordion.stories.tsx
@@ -117,6 +117,9 @@ const meta = {
   tags: ['autodocs'],
   excludeStories: /.*ArgsData$/,
   parameters: {
+    c1Meta: {
+      classification: ['requires-adapter'],
+    },
     layout: 'padded',
   },
   args: {